+++ tests/testthat/test_plot_KDE.R
@@ -14,6 +14,8 @@ test_that("input validation", {
   expect_error(expect_warning(plot_KDE(data.frame(a = Inf, b = 1)),
                               "Inf values removed in rows: 1 in data.frame 1"),
                "Your input is empty due to Inf removal")
+  expect_error(plot_KDE(df, df),
+               "'na.rm' should be a single logical value")
   expect_error(plot_KDE(df, summary.method = "error"),
                "'summary.method' should be one of 'MCM', 'weighted' or 'unweighted'")
   expect_error(plot_KDE(df, summary = 5),
@@ -24,6 +26,10 @@ test_that("input validation", {
                "'summary.pos' should be one of 'sub', 'left', 'center', 'right'")
   expect_error(plot_KDE(df, summary.pos = "error"),
                "'summary.pos' should be one of 'sub', 'left', 'center', 'right'")
+  expect_error(plot_KDE(df, bw = df),
+               "'bw' should be of class 'character' or 'numeric'")
+  expect_error(plot_KDE(df, bw = c(0, 1)),
+               "'bw' should be a single positive value")
   expect_error(plot_KDE(df, ylim = c(0, 1)),
                "'ylim' should have length 4")